py-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
numpy==1.24.3
pandas==2.0.3
